requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astropy==5.2.1
astroquery @ git+https://github.com/astropy/astroquery.git@03c8420cfee3d6c6654abfb63f85fdcc3d0cd686
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
binaryornot==0.4.4
bleach==6.0.0
bordercrop==1.0.0
CD==1.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.39.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jinja2-time==0.2.0
json-tricks==3.16.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.13.1
kiwisolver==1.4.4
lmdb==1.4.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
monai==1.1.0
more-itertools==9.1.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
ninja==1.11.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
orderedmultidict==1.0.1
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pathtools==0.1.2
pexpect==4.8.0
photutils==1.6.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.3.1
pyerfa==2.0.0.1
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2022.7.1
pyvo==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.10.1
SecretStorage==3.3.3
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
simplejson==3.18.4
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
src @ file:///home/dguthruf/Dev/astro-deconv
stack-data==0.6.2
tabulate==0.9.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchgeometry==0.1.2
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==3.0.1
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
wandb==0.13.11
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.5
zipp==3.15.0
